Fixing warnings

+++ str.h
@@ -52,7 +52,7 @@ static ssize_t
 Strtok(String string, String *out, char c)
 {
 	char *tmpptr = string.data;
-	while ((tmpptr - string.data) < string.len) {
+	while ((unsigned)(tmpptr - string.data) < string.len) {
 		if (*(++tmpptr) == c) break;
 	}
 	string.len = tmpptr - string.data;
@@ -63,7 +63,7 @@ Strtok(String string, String *out, char c)
 static String
 Strtrim(String str)
 {
-	int i = 0;
+	size_t i = 0;
 	while (isspace(*(str.data)) && i < str.len) {
 		++str.data;
 		--str.len;
